+++ bins/crm-chat-web/tests/e2e-telegram/media-visual.spec.ts
@@ -64,7 +64,7 @@ test.describe("Media Visual — Real Telegram Data", () => {
 
     const robot = await getRobotClient(workerCfg);
     registeredClientId = (await robot.mutation(
-      api.clients.workerRegisterConnected,
+      api.model.clients.workerRegisterConnected,
       {
         userId: convexUserId,
         telegramId,
